chore: component locations, minor improvements
