+++ crates/pop-common/src/lib.rs
@@ -23,3 +23,50 @@ impl Status for () {
 	// no-op: status updates are ignored
 	fn update(&self, _: &str) {}
 }
+
+/// Determines the target triple based on the current platform.
+pub fn target() -> Result<&'static str, Error> {
+	use std::env::consts::*;
+
+	if OS == "windows" {
+		return Err(Error::UnsupportedPlatform { arch: ARCH, os: OS });
+	}
+
+	match ARCH {
+		"aarch64" => {
+			return match OS {
+				"macos" => Ok("aarch64-apple-darwin"),
+				_ => Ok("aarch64-unknown-linux-gnu"),
+			}
+		},
+		"x86_64" | "x86" => {
+			return match OS {
+				"macos" => Ok("x86_64-apple-darwin"),
+				_ => Ok("x86_64-unknown-linux-gnu"),
+			}
+		},
+		&_ => {},
+	}
+	Err(Error::UnsupportedPlatform { arch: ARCH, os: OS })
+}
+
+#[cfg(test)]
+mod test {
+	use super::*;
+	use anyhow::Result;
+
+	#[test]
+	fn target_works() -> Result<()> {
+		use std::{process::Command, str};
+		let output = Command::new("rustc").arg("-vV").output()?;
+		let output = str::from_utf8(&output.stdout)?;
+		let target_expected = output
+			.lines()
+			.find(|l| l.starts_with("host: "))
+			.map(|l| &l[6..])
+			.unwrap()
+			.to_string();
+		assert_eq!(target()?, target_expected);
+		Ok(())
+	}
+}
